break layout/draft.js into smaller files
