+++ sites/shared/components/workbench/draft/utils.js
@@ -34,3 +34,13 @@ export const getProps = (obj) => {
 
   return props
 }
+
+export const dx = (pointA, pointB) => pointB.x - pointA.x
+export const dy = (pointA, pointB) => pointB.y - pointA.y
+export const rad2deg = radians => radians * 57.29577951308232
+export const angle = (pointA, pointB) => {
+  let rad = Math.atan2(-1 * dy(pointA, pointB), dx(pointA, pointB))
+  while (rad < 0) rad += 2 * Math.PI
+
+  return rad2deg(rad)
+}
